统计方形

+++ src/P2241_统计方形.java
@@ -5,12 +5,14 @@ public class P2241_统计方形 {
 		Scanner sc = new Scanner(System.in);
 		long n = sc.nextLong();
 		long m = sc.nextLong();
-		int cntSquare = 0;
-		int cntRectangle = 0;
+		long cntSquare = 0;
+		long cntRectangle = 0;
 
-		for (int i = 0; i < n; i++) {
-			for (int j = 0; j < m; j++) {
-				long cnt = (n - i) * (m - j);
+		for (long i = 0; i < n; i++) {
+			for (long j = 0; j < m; j++) {
+//				BigInteger cnt =
+//					(new BigInteger(String.valueOf(n)).add(new BigInteger(String.valueOf(-i)))).multiply((new BigInteger(String.valueOf(m)).add(new BigInteger(String.valueOf(-j)))))
+					long cnt = (n - i) * (m - j);
 				if (i == j) {
 					cntSquare += cnt;
 				} else {
@@ -18,6 +20,7 @@ public class P2241_统计方形 {
 				}
 			}
 		}
+
 		System.out.printf("%d %d", cntSquare, cntRectangle);
 	}
 }
